Nix coroutine, instead track distance we need to "use-up".
This is probably the most correct as everything else works by distance.
When using the coroutine it would wait time + inter-frame delays.

But, when I first tried this it was also "seeming to lag". The problem
was that I was still spawning the new sections 1 unit back, so waiting
to move them either way was causing them to end-up 2 units back.

So this works. I could go back to the coroutine and do the time-based
solution and it'd probably work better now. Plus it might be better
because there'd be less to do on "waiting" segments in the main loop.
But I'll keep it this way because of the aforementioned accuracy.