Made the Worm a prefab.
I was going to spawn a new one when you died, but on second thought I'll
keep the same head and just move it to a new location. That way I don't
need to revise the camera constraints or update any other references to
the worm head.

It's still a prefab... eventually there may be other worms to wriggle
alongside.