Sorted out the problem with rapid dying.
Again, it has to do with rapidly adding new segments. We normally only
suppress self-collisions with the first segment after the head, but if
we started out as a head and immediately added 2 segments, they'd both
be on the head location while they used up their "wait distance".

So: ignore any collisions we have with "waiting" segments.