Add a (very simple for now) AI to steer another worm.
I'm doing something a little weird which is keeping the Worm class
exactly the same for every worm, but then there is a "Steering" class
that should be added as a component before it. The steering class will
have Update() called first and should set steering properties in the
worm.

I can't reorder components in a prefab instance though, so in the prefab
I added BOTH steering components, and then one of them needs to be
disabled in any particular instance.

Perhaps instead I should just extend Worm, but I'm going to run with
these scissors for a little while.